访问者接口由any调整为范型，确保类型安全。避免指针和结构体使用错误时编译期无法发现错误。又由于go的范型无法只在方法上定义，所以把访问方法从表达式中剥离

+++ env/def_env_test.go
@@ -19,11 +19,11 @@ func TestBaseEnvironment_Get_Put(t *testing.T) {
 	env.PutDouble("testDouble", 3.14159)
 
 	assert.Equal(t, "bar", env.Get("foo").GetValue())
-	assert.Equal(t, int32(42), env.Get("baz").GetValue())
+	assert.Equal(t, 42, env.Get("baz").GetValue())
 	assert.Equal(t, true, env.Get("qux").GetValue())
 	assert.Equal(t, 3.14, env.Get("quux").GetValue())
 	assert.Equal(t, "Hello, World!", env.Get("testString").AsString())
-	assert.Equal(t, int32(100), env.Get("testInt").AsInteger())
+	assert.Equal(t, 100, int(env.Get("testInt").AsInteger()))
 	assert.Equal(t, true, env.Get("testBool").AsBoolean())
 	assert.Equal(t, 3.14159, env.Get("testDouble").AsDouble())
 	assert.Equal(t, 8, env.Size(), "环境变量数量应为8")
